docs: updating announcement (#2423)
