+++ .changeset/empty-donkeys-wash.md
@@ -1,0 +1,5 @@
+---
+'@module-federation/native-federation-typescript': patch
+---
+
+rollback without piscina
