+++ apps/website-new/docs/en/blog/announcement.mdx
@@ -3,8 +3,9 @@
 > April 26, 2024
 
 Today, we are excited to announce the official release of Module Federation 2.0!
-This new version is a collaborative effort between the [ByteDance Web Infra](https://github.com/web-infra-dev) team and the author of Module Federation, [Zack Jackson](https://github.com/ScriptedAlchemy).
-Module Federation 2.0 addresses many issues from the previous version, building upon Bytedance's internal practices and the existing community ecosystem of Module Federation.
+This new version is a collaborative effort between the [ByteDance Web Infra](https://github.com/web-infra-dev) team, the author of Module Federation, [Zack Jackson](https://github.com/ScriptedAlchemy)
+and several of our community members. Special thanks to [@ilteoood](https://github.com/ilteoood), [@patricklafrance](https://github.com/patricklafrance) and [@crutch12](https://github.com/crutch12).
+Module Federation 2.0 addresses many issues from the previous version, building upon ByteDance's internal practices and the existing community ecosystem of Module Federation.
 
 
 import AnnouncementVideo from '@components/AnnouncementVideo.tsx';
