+++ apps/website-new/docs/zh/blog/announcement.mdx
@@ -2,7 +2,8 @@
 
 > April 26, 2024
 
-今天，我们很高兴地宣布 Module Federation 2.0 正式发布了！这是一个由 [ByteDance Web Infra](https://github.com/web-infra-dev) 团队与 Module Federation 的作者 [@Zack Jackson](https://github.com/ScriptedAlchemy) 共同推出的新版本。Module Federation 2.0 解决了原有版本存在的诸多问题，新版本基于 Bytedance 内部实践以及 Module Federation 原有社区生态实现的。
+今天，我们很高兴地宣布 Module Federation 2.0 正式发布了！这是一个由 [ByteDance Web Infra](https://github.com/web-infra-dev) 团队、 Module Federation 的作者 [@Zack Jackson](https://github.com/ScriptedAlchemy) 社区成员共同推出的新版本。特别感谢 [@ilteoood](https://github.com/ilteoood)、[@patricklafrance](https://github.com/patricklafrance) 和 [@crutch12](https://github.com/crutch12)。
+Module Federation 2.0 解决了原有版本存在的诸多问题，新版本基于 Bytedance 内部实践以及 Module Federation 原有社区生态实现的。
 
 
 import AnnouncementVideo from '@components/AnnouncementVideo.tsx';
