+++ packages/native-federation-typescript/src/index.ts
@@ -1,4 +1,5 @@
 import ansiColors from 'ansi-colors';
+import { rm } from 'fs/promises';
 import { resolve } from 'path';
 import { mergeDeepRight } from 'rambda';
 import { UnpluginOptions, createUnplugin } from 'unplugin';
@@ -7,22 +8,31 @@ import { retrieveHostConfig } from './configurations/hostPlugin';
 import { retrieveRemoteConfig } from './configurations/remotePlugin';
 import { HostOptions } from './interfaces/HostOptions';
 import { RemoteOptions } from './interfaces/RemoteOptions';
-import { retrieveOriginalOutDir } from './lib/typeScriptCompiler';
-// @ts-expect-error function exported through module.exports
-import remoteWriteBundle from './lib/writeBundle/remote';
-// @ts-expect-error function exported through module.exports
-import hostWriteBundle from './lib/writeBundle/host';
+import { createTypesArchive, downloadTypesArchive } from './lib/archiveHandler';
+import {
+  compileTs,
+  retrieveMfTypesPath,
+  retrieveOriginalOutDir,
+} from './lib/typeScriptCompiler';
 
 export const NativeFederationTypeScriptRemote = createUnplugin(
   (options: RemoteOptions) => {
-    const remoteConfig = retrieveRemoteConfig(options);
-    const { remoteOptions, tsConfig } = remoteConfig;
+    const { remoteOptions, tsConfig, mapComponentsToExpose } =
+      retrieveRemoteConfig(options);
     return {
       name: 'native-federation-typescript/remote',
       async writeBundle() {
         try {
-          await remoteWriteBundle(remoteConfig);
+          compileTs(mapComponentsToExpose, tsConfig, remoteOptions);
 
+          await createTypesArchive(tsConfig, remoteOptions);
+
+          if (remoteOptions.deleteTypesFolder) {
+            await rm(retrieveMfTypesPath(tsConfig, remoteOptions), {
+              recursive: true,
+              force: true,
+            });
+          }
           console.log(ansiColors.green('Federated types created correctly'));
         } catch (error) {
           console.error(
@@ -68,12 +78,26 @@ export const NativeFederationTypeScriptRemote = createUnplugin(
 
 export const NativeFederationTypeScriptHost = createUnplugin(
   (options: HostOptions) => {
-    const hostConfig = retrieveHostConfig(options);
+    const { hostOptions, mapRemotesToDownload } = retrieveHostConfig(options);
     return {
       name: 'native-federation-typescript/host',
       async writeBundle() {
-        await hostWriteBundle(hostConfig);
+        if (hostOptions.deleteTypesFolder) {
+          await rm(hostOptions.typesFolder, {
+            recursive: true,
+            force: true,
+          }).catch((error) =>
+            console.error(
+              ansiColors.red(`Unable to remove types folder, ${error}`),
+            ),
+          );
+        }
 
+        const typesDownloader = downloadTypesArchive(hostOptions);
+        const downloadPromises =
+          Object.entries(mapRemotesToDownload).map(typesDownloader);
+
+        await Promise.allSettled(downloadPromises);
         console.log(ansiColors.green('Federated types extraction completed'));
       },
       get vite() {
